requirements.txt
aniso8601==9.0.1
boto3==1.26.45
botocore==1.29.45
cachetools==5.3.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
cv==1.0.0
decorator==4.4.2
Flask==2.2.2
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
google-api-core==2.11.0
google-api-python-client==2.74.0
google-auth==2.16.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.8.0
googleapis-common-protos==1.58.0
gunicorn==20.0.4
httplib2==0.21.0
idna==3.4
imageio==2.25.1
imageio-ffmpeg==0.4.8
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
keyboard==0.13.5
MarkupSafe==2.1.2
moviepy==1.0.3
numpy==1.23.4
oauthlib==3.2.2
opencv-python==4.7.0.72
Pillow==9.4.0
proglog==0.1.10
protobuf==4.21.12
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
pyserial==3.5
python-dateutil==2.8.2
python-dotenv==0.21.1
pytube==12.1.2
pytz==2022.7.1
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
s3transfer==0.6.0
six==1.16.0
tqdm==4.64.1
uritemplate==4.1.1
urllib3==1.26.12
Werkzeug==2.2.2
